+++ skforecast/ForecasterAutoreg/tests/test_set_lags.py
@@ -42,3 +42,23 @@ def test_set_lags_when_lags_argument_is_1d_numpy_array():
 
     assert (forecaster.lags == np.array([1, 2, 3])).all()
     assert forecaster.max_lag == 3
+
+
+def test_set_lags_when_differentiation_is_not_None():
+    """
+    Test how `window_size` is also updated when the forecaster includes differentiation.
+    """
+    lags = 5
+    differentiation = 1
+
+    forecaster = ForecasterAutoreg(
+                     regressor       = LinearRegression(),
+                     lags            = lags,
+                     differentiation = differentiation
+                 )
+    
+    forecaster.set_lags(lags=lags)
+
+    assert (forecaster.lags == np.array([1, 2, 3, 4, 5])).all()
+    assert forecaster.max_lag == 5
+    assert forecaster.window_size == lags + differentiation