+++ skforecast/ForecasterAutoregCustom/ForecasterAutoregCustom.py
@@ -1,8 +1,7 @@
 ################################################################################
 #                        ForecasterAutoregCustom                               #
 #                                                                              #
-# This work by Joaquin Amat Rodrigo and Javier Escobar Ortiz is licensed       #
-# under the BSD 3-Clause License.                                              #
+# This work by skforecast team is licensed under the BSD 3-Clause License.     #
 ################################################################################
 # coding=utf-8
 
@@ -231,16 +230,18 @@ class ForecasterAutoregCustom(ForecasterBase):
         self.python_version             = sys.version.split(" ")[0]
         self.forecaster_id              = forecaster_id
         
-        if not isinstance(window_size, int):
-            raise TypeError(
-                f"Argument `window_size` must be an int. Got {type(window_size)}."
+        if not isinstance(window_size, int) or window_size < 1:
+            raise ValueError(
+                (f"Argument `window_size` must be an integer equal to or "
+                 f"greater than 1. Got {window_size}.")
             )
         
-        if differentiation is not None and differentiation < 1:
-            raise ValueError(
-                f"`differentiation` must be greater than 0. Got {differentiation}."
-            )
         if self.differentiation is not None:
+            if not isinstance(differentiation, int) or differentiation < 1:
+                raise ValueError(
+                    (f"Argument `differentiation` must be an integer equal to or "
+                     f"greater than 1. Got {differentiation}.")
+                )
             self.window_size += self.differentiation
             self.differentiator = TimeSeriesDifferentiator(order=self.differentiation)
 
@@ -663,7 +664,9 @@ class ForecasterAutoregCustom(ForecasterBase):
         """
 
         if last_window is None:
-            last_window = copy(self.last_window)
+            last_window = self.last_window.copy()
+        else:
+            last_window = last_window.copy()
 
         last_window = last_window.iloc[-self.window_size:]
 
@@ -775,15 +778,14 @@ class ForecasterAutoregCustom(ForecasterBase):
         exog : pandas Series, pandas DataFrame, default `None`
             Exogenous variable/s included as predictor/s.
         n_boot : int, default `500`
-            Number of bootstrapping iterations used to estimate prediction
-            intervals.
+            Number of bootstrapping iterations used to estimate predictions.
         random_state : int, default `123`
-            Sets a seed to the random generator, so that boot intervals are always 
+            Sets a seed to the random generator, so that boot predictions are always 
             deterministic.
         in_sample_residuals : bool, default `True`
             If `True`, residuals from the training data are used as proxy of
-            prediction error to create prediction intervals. If `False`, out of
-            sample residuals are used. In the latter case, the user should have
+            prediction error to create predictions. If `False`, out of sample 
+            residuals are used. In the latter case, the user should have
             calculated and stored the residuals within the forecaster (see
             `set_out_sample_residuals()`).
 
@@ -810,7 +812,9 @@ class ForecasterAutoregCustom(ForecasterBase):
             )
 
         if last_window is None:
-            last_window = copy(self.last_window)
+            last_window = self.last_window.copy()
+        else:
+            last_window = last_window.copy()
 
         check_predict_input(
             forecaster_name  = type(self).__name__,
@@ -963,15 +967,14 @@ class ForecasterAutoregCustom(ForecasterBase):
             percentiles to compute, which must be between 0 and 100 inclusive. 
             For example, interval of 95% should be as `interval = [2.5, 97.5]`.
         n_boot : int, default `500`
-            Number of bootstrapping iterations used to estimate prediction
-            intervals.
+            Number of bootstrapping iterations used to estimate predictions.
         random_state : int, default `123`
-            Sets a seed to the random generator, so that boot intervals are always 
+            Sets a seed to the random generator, so that boot predictions are always 
             deterministic.
         in_sample_residuals : bool, default `True`
             If `True`, residuals from the training data are used as proxy of
-            prediction error to create prediction intervals. If `False`, out of
-            sample residuals are used. In the latter case, the user should have
+            prediction error to create predictions. If `False`, out of sample 
+            residuals are used. In the latter case, the user should have
             calculated and stored the residuals within the forecaster (see
             `set_out_sample_residuals()`).
 
@@ -1018,6 +1021,79 @@ class ForecasterAutoregCustom(ForecasterBase):
         return predictions
 
 
+    def predict_quantiles(
+        self,
+        steps: int,
+        last_window: Optional[pd.Series]=None,
+        exog: Optional[Union[pd.Series, pd.DataFrame]]=None,
+        quantiles: list=[0.05, 0.5, 0.95],
+        n_boot: int=500,
+        random_state: int=123,
+        in_sample_residuals: bool=True
+    ) -> pd.DataFrame:
+        """
+        Calculate the specified quantiles for each step. After generating 
+        multiple forecasting predictions through a bootstrapping process, each 
+        quantile is calculated for each step.
+        
+        Parameters
+        ----------
+        steps : int
+            Number of future steps predicted.
+        last_window : pandas Series, default `None`
+            Series values used to create the predictors (lags) needed in the 
+            first iteration of the prediction (t + 1).
+            If `last_window = None`, the values stored in` self.last_window` are
+            used to calculate the initial predictors, and the predictions start
+            right after training data.
+        exog : pandas Series, pandas DataFrame, default `None`
+            Exogenous variable/s included as predictor/s.
+        quantiles : list, default `[0.05, 0.5, 0.95]`
+            Sequence of quantiles to compute, which must be between 0 and 1 
+            inclusive. For example, quantiles of 0.05, 0.5 and 0.95 should be as 
+            `quantiles = [0.05, 0.5, 0.95]`.
+        n_boot : int, default `500`
+            Number of bootstrapping iterations used to estimate quantiles.
+        random_state : int, default `123`
+            Sets a seed to the random generator, so that boot quantiles are always 
+            deterministic.
+        in_sample_residuals : bool, default `True`
+            If `True`, residuals from the training data are used as proxy of
+            prediction error to create prediction quantiles. If `False`, out of
+            sample residuals are used. In the latter case, the user should have
+            calculated and stored the residuals within the forecaster (see
+            `set_out_sample_residuals()`).
+
+        Returns
+        -------
+        predictions : pandas DataFrame
+            Quantiles predicted by the forecaster.
+
+        Notes
+        -----
+        More information about prediction intervals in forecasting:
+        https://otexts.com/fpp2/prediction-intervals.html
+        Forecasting: Principles and Practice (2nd ed) Rob J Hyndman and
+        George Athanasopoulos.
+        
+        """
+        
+        check_interval(quantiles=quantiles)
+
+        boot_predictions = self.predict_bootstrapping(
+                               steps               = steps,
+                               last_window         = last_window,
+                               exog                = exog,
+                               n_boot              = n_boot,
+                               random_state        = random_state,
+                               in_sample_residuals = in_sample_residuals
+                           )
+
+        predictions = boot_predictions.quantile(q=quantiles, axis=1).transpose()
+
+        return predictions
+
+
     def predict_dist(
         self,
         steps: int,
@@ -1048,15 +1124,14 @@ class ForecasterAutoregCustom(ForecasterBase):
         exog : pandas Series, pandas DataFrame, default `None`
             Exogenous variable/s included as predictor/s.
         n_boot : int, default `500`
-            Number of bootstrapping iterations used to estimate prediction
-            intervals.
+            Number of bootstrapping iterations used to estimate predictions.
         random_state : int, default `123`
-            Sets a seed to the random generator, so that boot intervals are always 
+            Sets a seed to the random generator, so that boot predictions are always 
             deterministic.
         in_sample_residuals : bool, default `True`
             If `True`, residuals from the training data are used as proxy of
-            prediction error to create prediction intervals. If `False`, out of
-            sample residuals are used. In the latter case, the user should have
+            prediction error to create predictions. If `False`, out of sample 
+            residuals are used. In the latter case, the user should have
             calculated and stored the residuals within the forecaster (see
             `set_out_sample_residuals()`).
 
@@ -1079,9 +1154,9 @@ class ForecasterAutoregCustom(ForecasterBase):
         param_names = [p for p in inspect.signature(distribution._pdf).parameters
                        if not p=='x'] + ["loc","scale"]
         param_values = np.apply_along_axis(
-                            lambda x: distribution.fit(x),
-                            axis = 1,
-                            arr  = boot_samples
+                           lambda x: distribution.fit(x),
+                           axis = 1,
+                           arr  = boot_samples
                        )
         predictions = pd.DataFrame(
                           data    = param_values,