+++ skforecast/ForecasterAutoregMultiSeries/tests/test_predict_quantiles.py
@@ -1,0 +1,95 @@
+# Unit test predict_quantiles ForecasterAutoregMultiSeries
+# ==============================================================================
+import pytest
+import numpy as np
+import pandas as pd
+from skforecast.ForecasterAutoregMultiSeries import ForecasterAutoregMultiSeries
+from sklearn.compose import ColumnTransformer
+from sklearn.preprocessing import StandardScaler
+from sklearn.preprocessing import OneHotEncoder
+from sklearn.linear_model import LinearRegression
+
+# Fixtures
+from .fixtures_ForecasterAutoregMultiSeries import series
+from .fixtures_ForecasterAutoregMultiSeries import exog
+from .fixtures_ForecasterAutoregMultiSeries import exog_predict
+
+transformer_exog = ColumnTransformer(
+                       [('scale', StandardScaler(), ['exog_1']),
+                        ('onehot', OneHotEncoder(), ['exog_2'])],
+                       remainder = 'passthrough',
+                       verbose_feature_names_out = False
+                   )
+
+
+@pytest.mark.parametrize("level", 
+                         ['1', ['1']], 
+                         ids=lambda lvl: f'level: {lvl}')
+def test_predict_quantiles_output_when_forecaster_is_LinearRegression_steps_is_2_in_sample_residuals_True_exog_and_transformer(level):
+    """
+    Test output of predict_quantiles when regressor is LinearRegression,
+    2 steps are predicted, using in-sample residuals, exog is included and both
+    inputs are transformed. Single level.
+    """
+    forecaster = ForecasterAutoregMultiSeries(
+                     regressor          = LinearRegression(),
+                     lags               = 3,
+                     transformer_series = StandardScaler(),
+                     transformer_exog   = transformer_exog,
+                 )
+    
+    forecaster.fit(series=series, exog=exog)
+    results = forecaster.predict_quantiles(
+                  steps               = 2,
+                  quantiles           = [0.05, 0.55, 0.95],
+                  levels              = level,
+                  exog                = exog_predict,
+                  n_boot              = 4,
+                  in_sample_residuals = True
+              )
+    
+    expected = pd.DataFrame(
+                   data    = np.array([[0.20133572, 0.40460928, 0.53761911],
+                                       [0.07739771, 0.19454095, 0.45374409]]),
+                   columns = ['1_0.05', '1_0.55', '1_0.95'],
+                   index   = pd.RangeIndex(start=50, stop=52)
+               )
+    
+    pd.testing.assert_frame_equal(results, expected)
+
+
+@pytest.mark.parametrize("levels", 
+                         [['1', '2'], None], 
+                         ids=lambda lvl: f'levels: {lvl}')
+def test_predict_quantiles_output_when_forecaster_is_LinearRegression_steps_is_2_in_sample_residuals_False_exog_and_transformer(levels):
+    """
+    Test output of predict_quantiles when regressor is LinearRegression,
+    2 steps are predicted, using out-sample residuals, exog is included and both
+    inputs are transformed. Multiple levels.
+    """
+    forecaster = ForecasterAutoregMultiSeries(
+                     regressor          = LinearRegression(),
+                     lags               = 3,
+                     transformer_series = StandardScaler(),
+                     transformer_exog   = transformer_exog,
+                 )
+    
+    forecaster.fit(series=series, exog=exog)
+    forecaster.out_sample_residuals = forecaster.in_sample_residuals
+    results = forecaster.predict_quantiles(
+                  steps               = 2,
+                  quantiles           = [0.05, 0.55, 0.95],
+                  levels              = levels,
+                  exog                = exog_predict,
+                  n_boot              = 4,
+                  in_sample_residuals = False
+              )
+    
+    expected = pd.DataFrame(
+                   data    = np.array([[0.20133572, 0.40460928, 0.53761911, 0.27660015, 0.61593304, 0.7616801 ],
+                                       [0.07739771, 0.19454095, 0.45374409, 0.18601931, 0.40403914, 0.70461714]]),
+                   columns = ['1_0.05', '1_0.55', '1_0.95', '2_0.05', '2_0.55', '2_0.95'],
+                   index   = pd.RangeIndex(start=50, stop=52)
+               )
+
+    pd.testing.assert_frame_equal(results, expected)