+++ skforecast/ForecasterAutoregMultiSeriesCustom/tests/test_predict_bootstrapping.py
@@ -17,8 +17,8 @@ from .fixtures_ForecasterAutoregMultiSeriesCustom import exog
 from .fixtures_ForecasterAutoregMultiSeriesCustom import exog_predict
 
 transformer_exog = ColumnTransformer(
-                       [('scale', StandardScaler(), ['col_1']),
-                        ('onehot', OneHotEncoder(), ['col_2'])],
+                       [('scale', StandardScaler(), ['exog_1']),
+                        ('onehot', OneHotEncoder(), ['exog_2'])],
                        remainder = 'passthrough',
                        verbose_feature_names_out = False
                    )
@@ -154,8 +154,8 @@ def test_predict_bootstrapping_output_when_forecaster_is_LinearRegression_exog_s
                      fun_predictors  = create_predictors,
                      window_size     = 3
                  )
-    forecaster.fit(series=series, exog=exog['col_1'])
-    results = forecaster.predict_bootstrapping(steps=1, n_boot=4, exog=exog_predict['col_1'], in_sample_residuals=True)
+    forecaster.fit(series=series, exog=exog['exog_1'])
+    results = forecaster.predict_bootstrapping(steps=1, n_boot=4, exog=exog_predict['exog_1'], in_sample_residuals=True)
 
     expected_1 = pd.DataFrame(
                      data    = np.array([[0.16386591, 0.56398143, 0.38576231, 0.37782729]]),
@@ -184,8 +184,8 @@ def test_predict_bootstrapping_output_when_forecaster_is_LinearRegression_exog_s
                      fun_predictors  = create_predictors,
                      window_size     = 3
                  )
-    forecaster.fit(series=series, exog=exog['col_1'])
-    results = forecaster.predict_bootstrapping(steps=2, n_boot=4, exog=exog_predict['col_1'], in_sample_residuals=True)
+    forecaster.fit(series=series, exog=exog['exog_1'])
+    results = forecaster.predict_bootstrapping(steps=2, n_boot=4, exog=exog_predict['exog_1'], in_sample_residuals=True)
 
     expected_1 = pd.DataFrame(
                      data    = np.array([[0.16386591, 0.56398143, 0.38576231, 0.37782729],
@@ -216,9 +216,9 @@ def test_predict_bootstrapping_output_when_forecaster_is_LinearRegression_exog_s
                      fun_predictors  = create_predictors,
                      window_size     = 3
                  )
-    forecaster.fit(series=series, exog=exog['col_1'])
+    forecaster.fit(series=series, exog=exog['exog_1'])
     forecaster.out_sample_residuals = forecaster.in_sample_residuals
-    results = forecaster.predict_bootstrapping(steps=1, n_boot=4, exog=exog_predict['col_1'], in_sample_residuals=False)
+    results = forecaster.predict_bootstrapping(steps=1, n_boot=4, exog=exog_predict['exog_1'], in_sample_residuals=False)
 
     expected_1 = pd.DataFrame(
                      data    = np.array([[0.16386591, 0.56398143, 0.38576231, 0.37782729]]),
@@ -247,9 +247,9 @@ def test_predict_bootstrapping_output_when_forecaster_is_LinearRegression_exog_s
                      fun_predictors  = create_predictors,
                      window_size     = 3
                  )
-    forecaster.fit(series=series, exog=exog['col_1'])
+    forecaster.fit(series=series, exog=exog['exog_1'])
     forecaster.out_sample_residuals = forecaster.in_sample_residuals
-    results = forecaster.predict_bootstrapping(steps=2, n_boot=4, exog=exog_predict['col_1'], in_sample_residuals=False)
+    results = forecaster.predict_bootstrapping(steps=2, n_boot=4, exog=exog_predict['exog_1'], in_sample_residuals=False)
 
     expected_1 = pd.DataFrame(
                      data    = np.array([[0.16386591, 0.56398143, 0.38576231, 0.37782729],