+++ skforecast/plot/plot.py
@@ -1,8 +1,7 @@
 ################################################################################
 #                                   Plot                                       #
 #                                                                              #
-# This work by Joaquin Amat Rodrigo and Javier Escobar Ortiz is licensed       #
-# under the BSD 3-Clause License.                                              #
+# This work by skforecast team is licensed under the BSD 3-Clause License.     #
 ################################################################################
 # coding=utf-8