+++ skforecast/preprocessing/preprocessing.py
@@ -1,8 +1,7 @@
 ################################################################################
 #                         skforecast.preprocessing                             #
 #                                                                              #
-# This work by Joaquin Amat Rodrigo and Javier Escobar Ortiz is licensed       #
-# under the BSD 3-Clause License.                                              #
+# This work by skforecast team is licensed under the BSD 3-Clause License.     #
 ################################################################################
 # coding=utf-8
 
@@ -115,11 +114,7 @@ class TimeSeriesDifferentiator(BaseEstimator, TransformerMixin):
 
         """
         
-        for i in range(self.order):
-            if i == 0:
-                X_diff = np.diff(X, n=1)
-            else:
-                X_diff = np.diff(X_diff, n=1)
+        X_diff = np.diff(X, n=self.order)
                 
         X_diff = np.append((np.full(shape=self.order, fill_value=np.nan)), X_diff)